+++ src/MudBlazor/State/Comparer/ParameterHandlerUniquenessComparer.cs
@@ -1,19 +1,15 @@
-﻿// Copyright (c) MudBlazor 2021
-// MudBlazor licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
+﻿using System.Collections.Generic;
 
-using System.Collections.Generic;
-
-namespace MudBlazor.State;
+namespace MudBlazor.State.Comparer;
 
 #nullable enable
 /// <summary>
 /// Provides functionality to compare instances of <see cref="ParameterMetadata"/> or <see cref="IParameterComponentLifeCycle.Metadata"/> for equality.
 /// </summary>
 /// <remarks>
-/// This checks only uniqueness of <see cref="ParameterMetadata.ParameterName"/>.
+/// This checks only uniqueness of <see cref="ParameterMetadata.HandlerName"/>.
 /// </remarks>
-internal class ParameterNameUniquenessComparer : IEqualityComparer<ParameterMetadata>, IEqualityComparer<IParameterComponentLifeCycle>
+internal class ParameterHandlerUniquenessComparer : IEqualityComparer<ParameterMetadata>, IEqualityComparer<IParameterComponentLifeCycle>
 {
     /// <inheritdoc />
     public bool Equals(ParameterMetadata? x, ParameterMetadata? y)
@@ -33,7 +29,12 @@ internal class ParameterNameUniquenessComparer : IEqualityComparer<ParameterMeta
             return false;
         }
 
-        return x.ParameterName == y.ParameterName;
+        if (x.HandlerName is null && y.HandlerName is null)
+        {
+            return false;
+        }
+
+        return x.HandlerName == y.HandlerName;
     }
 
     /// <inheritdoc />
@@ -58,13 +59,18 @@ internal class ParameterNameUniquenessComparer : IEqualityComparer<ParameterMeta
     }
 
     /// <inheritdoc />
-    public int GetHashCode(ParameterMetadata parameterMetadata) => parameterMetadata.ParameterName.GetHashCode();
+    public int GetHashCode(ParameterMetadata parameterMetadata)
+    {
+        return parameterMetadata.HandlerName is not null
+            ? parameterMetadata.HandlerName.GetHashCode()
+            : parameterMetadata.GetHashCode();
+    }
 
     /// <inheritdoc />
     public int GetHashCode(IParameterComponentLifeCycle parameterComponentLifeCycle) => GetHashCode(parameterComponentLifeCycle.Metadata);
 
     /// <summary>
-    /// Gets the default instance of <see cref="ParameterNameUniquenessComparer"/>.
+    /// Gets the default instance of <see cref="ParameterHandlerUniquenessComparer"/>.
     /// </summary>
-    public static readonly ParameterNameUniquenessComparer Default = new();
+    public static readonly ParameterHandlerUniquenessComparer Default = new();
 }
