+++ src/MudBlazor/State/Comparer/ParameterNameUniquenessComparer.cs
@@ -1,15 +1,19 @@
-﻿using System.Collections.Generic;
+﻿// Copyright (c) MudBlazor 2021
+// MudBlazor licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
 
-namespace MudBlazor.State;
+using System.Collections.Generic;
+
+namespace MudBlazor.State.Comparer;
 
 #nullable enable
 /// <summary>
 /// Provides functionality to compare instances of <see cref="ParameterMetadata"/> or <see cref="IParameterComponentLifeCycle.Metadata"/> for equality.
 /// </summary>
 /// <remarks>
-/// This checks only uniqueness of <see cref="ParameterMetadata.HandlerName"/>.
+/// This checks only uniqueness of <see cref="ParameterMetadata.ParameterName"/>.
 /// </remarks>
-internal class ParameterHandlerUniquenessComparer : IEqualityComparer<ParameterMetadata>, IEqualityComparer<IParameterComponentLifeCycle>
+internal class ParameterNameUniquenessComparer : IEqualityComparer<ParameterMetadata>, IEqualityComparer<IParameterComponentLifeCycle>
 {
     /// <inheritdoc />
     public bool Equals(ParameterMetadata? x, ParameterMetadata? y)
@@ -29,12 +33,7 @@ internal class ParameterHandlerUniquenessComparer : IEqualityComparer<ParameterM
             return false;
         }
 
-        if (x.HandlerName is null && y.HandlerName is null)
-        {
-            return false;
-        }
-
-        return x.HandlerName == y.HandlerName;
+        return x.ParameterName == y.ParameterName;
     }
 
     /// <inheritdoc />
@@ -59,18 +58,13 @@ internal class ParameterHandlerUniquenessComparer : IEqualityComparer<ParameterM
     }
 
     /// <inheritdoc />
-    public int GetHashCode(ParameterMetadata parameterMetadata)
-    {
-        return parameterMetadata.HandlerName is not null
-            ? parameterMetadata.HandlerName.GetHashCode()
-            : parameterMetadata.GetHashCode();
-    }
+    public int GetHashCode(ParameterMetadata parameterMetadata) => parameterMetadata.ParameterName.GetHashCode();
 
     /// <inheritdoc />
     public int GetHashCode(IParameterComponentLifeCycle parameterComponentLifeCycle) => GetHashCode(parameterComponentLifeCycle.Metadata);
 
     /// <summary>
-    /// Gets the default instance of <see cref="ParameterHandlerUniquenessComparer"/>.
+    /// Gets the default instance of <see cref="ParameterNameUniquenessComparer"/>.
     /// </summary>
-    public static readonly ParameterHandlerUniquenessComparer Default = new();
+    public static readonly ParameterNameUniquenessComparer Default = new();
 }
