+++ src/MudBlazor/State/ParameterScopeContainer.cs
@@ -12,6 +12,7 @@ using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Components;
+using MudBlazor.State.Comparer;
 
 namespace MudBlazor.State;
